requirements.txt
# === Core Flask Framework ===
Flask==3.1.2
Flask-Login==0.6.3
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
blinker==1.9.0
click==8.2.1

# === Real-time Communication (WebSocket) ===
Flask-SocketIO==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2
bidict==0.23.1
simple-websocket==1.1.0

# === WSGI Production Server ===
gunicorn==23.0.0
eventlet==0.40.3
greenlet==3.2.4

# === Google Generative AI ===
google-generativeai==0.8.5
google-api-core==2.25.1
googleapis-common-protos==1.70.0
grpcio==1.75.0
protobuf==5.29.5
google-auth==2.20.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.87.0

# === OpenAI API (for AI Chat) ===
openai==1.58.1

# === Redis Cache ===
redis==5.2.1

# === External Data Services ===
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# === Environment Variables ===
python-dotenv==1.1.1

# === Utilities ===
pillow==11.3.0
numpy==1.26.4
cryptography==44.0.0
pandas==2.2.3

